[FEAT] 입력에 대한 예외상황 구현

+++ app/src/main/java/com/preonboarding/moviereview/data/remote/model/Review.kt
@@ -8,5 +8,5 @@ data class Review(
     val imageUrl: String,
     val nickName: String,
     val password: Int,
-    val star: Double,
+    val star: Float,
 )
